Safari-611.1.10/LayoutTests/http/tests/fileapi/xhr-send-form-data-mimetype-normalization-expected.txt

Test the normalization of Blob MIME types passed to FormData.append().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing the test infrastructure with a simple MIME type:
file=filename:text/plain:1234567890
PASS response is 'file=filename:text/plain:1234567890'
Sending FormData containing CRs in MIME type:
file=filename:application/octet-stream:1234567890
PASS response is 'file=filename:application/octet-stream:1234567890'
Sending FormData containing LF in MIME type:
file=filename:application/octet-stream:1234567890
PASS response is 'file=filename:application/octet-stream:1234567890'
Sending FormData containing a MIME header in its MIME type:
file=filename:application/octet-stream:1234567890&file2=filename2:text/plain:1234567890
PASS response is 'file=filename:application/octet-stream:1234567890&file2=filename2:text/plain:1234567890'
PASS successfullyParsed is true

TEST COMPLETE
